+++ src/server/stop.rs
@@ -2,7 +2,8 @@ use actix_web::{post, HttpResponse, Responder};
 use log::{info, trace};
 use std::process;
 
-use crate::util;
+use argon::util;
+// use argon::{core::Core, ext::ResultExt, server::AuthRequest}; // This line was added incorrectly and will be removed
 
 #[post("/stop")]
 async fn main() -> impl Responder {