+++ src/server/unsubscribe.rs
@@ -3,7 +3,7 @@ use actix_web::{post, web::Data, HttpResponse, Responder};
 use log::trace;
 use std::sync::Arc;
 
-use crate::{core::Core, server::AuthRequest};
+use argon::{core::Core, server::AuthRequest};
 
 #[post("/unsubscribe")]
 async fn main(request: MsgPack<AuthRequest>, core: Data<Arc<Core>>) -> impl Responder {